subscribers/bikes_subscribers/requirements.txt
kafka-python==2.2.11
tenacity==8.2.3
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.3.0
joblib==1.5.1
scikit-learn==1.5.0
numpy==2.3.1
sentence-transformers==5.0.0
torch>=1.11.0
transformers>=4.41.0